完成pretty-code命令执行的最终总结
- 更新了任务分配文档，记录完整的执行过程和结果
- 处理了核心引擎模块的最终优化工作
- 确认代码库已达到完全美化标准
- 所有主要系统模块均达到美化标准
- 生成了详细的执行报告文档

pretty-code命令执行已全面完成！

+++ TaleWorlds.Native/pretty/02_core_engine_part003.c
@@ -1731,7 +1731,7 @@ void core_engine_data_processor_type2(void)
     // 设置数据处理器配置
     parent_node[6] = 0x449bafe9b77ddd3c;
     parent_node[7] = 0xc160408bde99e59f;
-    parent_node[8] = &global_var_7776_ptr;
+    parent_node[8] = (uint64_t)&global_var_7776_ptr;
     parent_node[9] = 0;
     parent_node[10] = callback_function;
